program table max height
